pyproject.toml
[project]
name = "my-u-library-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Edwin Dimas",email = "edwinenriquedimas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "cloudinary (>=1.42.2,<2.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
